requirements.txt
fastmcp
httpx
python-dotenv
cachetools
# для mock-сервера
fastapi
uvicorn
